PDCL-6013 Organized identity map utils and constants. Removed authenticatedStateOptions.
