+++ src/view/dataElements/identityMap/utils/getDefaultIdentifier.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2020 Adobe. All rights reserved.
+Copyright 2019 Adobe. All rights reserved.
 This file is licensed to you under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License. You may obtain a copy
 of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -10,21 +10,12 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import fetchFromPlatform from "../../utils/fetchFromPlatform";
+import { AMBIGUOUS } from "../constants/authenticatedState";
 
-export default async ({ orgId, imsAccess }) => {
-  let parsedResponse;
-
-  try {
-    parsedResponse = await fetchFromPlatform({
-      orgId,
-      imsAccess,
-      path: "/data/core/idnamespace/identities"
-    });
-  } catch (e) {
-    // TODO: Should we be reporting an error instead?
-    return [];
-  }
-
-  return parsedResponse.parsedBody;
+export default () => {
+  return {
+    id: "",
+    authenticatedState: AMBIGUOUS,
+    primary: false
+  };
 };
