+++ src/view/dataElements/identityMap/utils/getDefaultIdentity.js
