--- src/view/utils/wrapValidateWithErrorLogging.js
@@ -1,33 +1,0 @@
-/*
-Copyright 2021 Adobe. All rights reserved.
-This file is licensed to you under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License. You may obtain a copy
-of the License at http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software distributed under
-the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
-OF ANY KIND, either express or implied. See the License for the specific language
-governing permissions and limitations under the License.
-*/
-
-// Wraps any validate function that will be called by Formik.
-// Without this, if an error is thrown during validation, Formik
-// swallows the error and it's difficult to figure out where the problem is.
-// https://github.com/jaredpalmer/formik/issues/1329
-export default validate => {
-  if (!validate) {
-    return undefined;
-  }
-
-  return (...args) => {
-    let result;
-    try {
-      result = validate(...args);
-    } catch (error) {
-      // eslint-disable-next-line no-console
-      console.error("Error executing validation", error);
-      throw error;
-    }
-    return result;
-  };
-};
